+++ docs/CHANGELOG.rst
@@ -14,6 +14,8 @@ Unreleased
 Added
 -----
 - Allow staff users to create new ``Variant`` objects through the API endpoint
+- Allow authenticated users to create new ``VariantCall`` objects via the API
+  endpoint
 
 Changed
 -------